Migrate to extensions
