+++ analysis_options.yaml
@@ -1,8 +1,3 @@
-analyzer:
-  enable-experiment:
-    - extension-methods
-#    - non-nullable
-
 # Lint rules and documentation, see http://dart-lang.github.io/linter/lints
 linter:
   rules:
