+++ example/main.dart
@@ -1,6 +1,4 @@
 import "package:kt_dart/kt.dart";
-import "package:kt_dart/src/collection/extension/iterable_extension_mixin.dart";
-//import "package:kt_dart/src/collection/extension/list_extension_mixin.dart";
 
 void main() {
   /// Lists
