+++ lib/src/collection/impl/list_empty.dart
@@ -1,7 +1,4 @@
 import "package:kt_dart/collection.dart";
-import "package:kt_dart/src/collection/extension/collection_extension_mixin.dart";
-import "package:kt_dart/src/collection/extension/iterable_extension_mixin.dart";
-import "package:kt_dart/src/collection/extension/list_extension_mixin.dart";
 import "package:kt_dart/src/collection/impl/dart_iterable.dart";
 
 class EmptyList<T> extends Object implements KtList<T> {
