+++ lib/src/collection/impl/map_empty.dart
@@ -1,5 +1,4 @@
 import "package:kt_dart/collection.dart";
-import "package:kt_dart/src/collection/extension/map_extensions_mixin.dart";
 
 class EmptyMap<K, V> extends Object implements KtMap<K, V> {
   EmptyMap();
