+++ lib/src/collection/impl/set.dart
@@ -1,6 +1,4 @@
 import "package:kt_dart/collection.dart";
-import "package:kt_dart/src/collection/extension/collection_extension_mixin.dart";
-import "package:kt_dart/src/collection/extension/iterable_extension_mixin.dart";
 import "package:kt_dart/src/collection/impl/dart_unmodifiable_set_view.dart";
 import "package:kt_dart/src/util/hash.dart";
 
@@ -62,6 +60,17 @@ class DartSet<T> extends Object implements KtSet<T> {
       return other.containsAll(this);
     }
     return false;
+  }
+
+  @override
+  String toString() {
+    return joinToString(
+      separator: ", ",
+      prefix: "[",
+      postfix: "]",
+      transform: (it) =>
+          identical(it, this) ? "(this Collection)" : it.toString(),
+    );
   }
 }
