+++ lib/src/collection/impl/set_mutable.dart
@@ -1,7 +1,4 @@
 import "package:kt_dart/collection.dart";
-import "package:kt_dart/src/collection/extension/collection_extension_mixin.dart";
-import "package:kt_dart/src/collection/extension/iterable_extension_mixin.dart";
-import "package:kt_dart/src/collection/extension/iterable_mutable_extension_mixin.dart";
 import "package:kt_dart/src/util/hash.dart";
 
 class DartMutableSet<T> extends Object implements KtMutableSet<T> {
@@ -108,6 +105,17 @@ class DartMutableSet<T> extends Object implements KtMutableSet<T> {
     final oldSize = size;
     _set.removeWhere((it) => !elements.contains(it));
     return oldSize != size;
+  }
+
+  @override
+  String toString() {
+    return joinToString(
+      separator: ", ",
+      prefix: "[",
+      postfix: "]",
+      transform: (it) =>
+          identical(it, this) ? "(this Collection)" : it.toString(),
+    );
   }
 }
